Groups tab now uses the new background color instead of stark bright white
